[refactor]-Add packages for types
