+++ umanz_app/pages/interne/conge/add.vue
@@ -6,7 +6,7 @@ definePageMeta({
 import { z } from 'zod'
 import { reactive, ref, computed } from 'vue'
 import type { FormConge } from '~/types/interne/dto';
-import type { TypeConge } from '~/types/interne';
+import type { TypeConge } from '~/types/interne/conge';
 
 const idContrat = computed(() => localStorage.getItem("umanz-idContrat"));
 const { demanderCongerFn, validateJustificatif } = useCongeActions();
