--- umanz_app/types/interne/dto.ts
@@ -1,12 +1,0 @@
-export interface JustificatifDTO {
-    dateJustificatif: string;
-    imageJustificatif: string;
-}
-
-export interface FormConge {
-    dateDebut: string;
-    nbJour: number;
-    motif: string;
-    idTypeConge: number;
-    justificatif: JustificatifDTO;
-}
