+++ umanz_app/types/interne/formation/index.ts
@@ -1,0 +1,6 @@
+import type { Contrat } from "../contrat";
+
+export interface Participation {
+    contrat: Contrat,
+    dateInscription: string
+}
